treshold

+++ r2/common.h
@@ -237,7 +237,7 @@ uniform sampler2D       s_tonemap;              // actually MidleGray / exp(Lw +
 #define def_hdr_clip	half(0.75h)        		//
 
 //////////////////////////////////////////////////////////////////////////////////////////
-#define	LUMINANCE_VECTOR                 half3(0.3f, 0.38f, 0.22f)
+#define	LUMINANCE_VECTOR                 float3(0.3f, 0.38f, 0.22f)
 void        tonemap              (out half4 low, out half4 high, half3 rgb, half scale)
 {
         rgb     =      	rgb*scale       ;
